Adding files and changing package heiarchy
I made a few changes to the heiarchy of packages. I was thinking that we put communicators in their own class,
Server stuff in it's own class, client stuff in it's own class, and things in common between the two in their
own class (i.e. Interfaces, DataModels, etc.). I have the ServerCommunicator working for the most part and a
few skeletons for other classes.