Fixed the issue with Command.
I made a dummy class for Command that only took in and stored the
parameters needed to make a Command object. That way, Command could
stay in the server package and get the ServerFacade functions while
also being "in" common. If Client needs to send a Command, they just
send a CommandParams object.