Fixed Socket issues. Added to AndroidManifest. Serialized Objects.
I changed the AndroidManifest so then we could allow Internet access
to the ServerSocket. Also added definition to the LobbyActivity so
we could switch to it from LoginActivity.
Serialized Objects in Common so then they could be sent over the
sockets. Apprently just serializing the parent class isn't enough.
When sending a class name, we need to send the "fully-qualified"
name of the object. Example: instead of "String", send "java.lang.String".
So with our classes in Common, we have to do "common.DataModels.the-class"